backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
reportlab==4.0.4
nltk==3.8.1
python-dotenv==1.0.0
gunicorn==21.2.0
